+++ src/types.ts
@@ -1,5 +1,6 @@
 import { ExtensionProvider } from '@elrondnetwork/erdjs-extension-provider';
 import { WalletConnectProvider } from '@elrondnetwork/erdjs-wallet-connect-provider';
+import { WalletProvider } from '@elrondnetwork/erdjs-web-wallet-provider/out';
 
 export interface InitOptions {
   apiUrl: string;
@@ -20,9 +21,11 @@ export enum LoginMethodsEnum {
 export type DappProvider =
   | ExtensionProvider
   | WalletConnectProvider
+  | WalletProvider
   | undefined;
 
 export interface LoginOptions {
   qrCodeContainer?: string | HTMLElement;
   token?: string;
+  callbackRoute?: string;
 }