+++ src/utils/constants.ts
@@ -1,4 +1,4 @@
-interface BaseNetworkType {
+interface NetworkType {
   id: string;
   shortId: string;
   name: string;
@@ -6,15 +6,10 @@ interface BaseNetworkType {
   egldDenomination: string;
   decimals: string;
   gasPerDataByte: string;
-  walletConnectDeepLink: string;
   walletAddress: string;
   apiAddress: string;
   explorerAddress: string;
   apiTimeout: number;
-}
-
-interface NetworkType extends BaseNetworkType {
-  walletConnectBridgeAddresses: string[];
 }
 
 export const LOCAL_STORAGE_KEY = 'elvenjs_state';
@@ -26,7 +21,13 @@ export const DEFAULT_MIN_GAS_LIMIT = 50_000;
 export const DAPP_CONFIG_ENDPOINT = '/dapp/config';
 export const DAPP_INIT_ROUTE = '/dapp/init';
 
-export const chainTypeConfig = 'devnet';
+export const defaultChainTypeConfig = 'devnet';
+
+export const walletConnectDeepLink =
+  'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://maiar.com/';
+export const walletConnectBridgeAddresses = [
+  'https://bridge.walletconnect.org',
+];
 
 export const networkConfig: Record<string, NetworkType> = {
   devnet: {
@@ -37,15 +38,11 @@ export const networkConfig: Record<string, NetworkType> = {
     egldDenomination: '18',
     decimals: '4',
     gasPerDataByte: '1500',
-    walletConnectDeepLink:
-      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://maiar.com/',
-    walletConnectBridgeAddresses: ['https://bridge.walletconnect.org'],
     walletAddress: 'https://devnet-wallet.elrond.com',
     apiAddress: 'https://devnet-api.elrond.com',
     explorerAddress: 'https://devnet-explorer.elrond.com',
     apiTimeout: 10000,
   },
-
   testnet: {
     id: 'testnet',
     shortId: 'T',
@@ -54,15 +51,11 @@ export const networkConfig: Record<string, NetworkType> = {
     egldDenomination: '18',
     decimals: '4',
     gasPerDataByte: '1500',
-    walletConnectDeepLink:
-      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://maiar.com/',
-    walletConnectBridgeAddresses: ['https://bridge.walletconnect.org'],
     walletAddress: 'https://testnet-wallet.elrond.com',
     apiAddress: 'https://testnet-api.elrond.com',
     explorerAddress: 'https://testnet-explorer.elrond.com',
     apiTimeout: 10000,
   },
-
   mainnet: {
     id: 'mainnet',
     shortId: '1',
@@ -71,9 +64,6 @@ export const networkConfig: Record<string, NetworkType> = {
     egldDenomination: '18',
     decimals: '4',
     gasPerDataByte: '1500',
-    walletConnectDeepLink:
-      'https://maiar.page.link/?apn=com.elrond.maiar.wallet&isi=1519405832&ibi=com.elrond.maiar.wallet&link=https://maiar.com/',
-    walletConnectBridgeAddresses: ['https://bridge.walletconnect.org'],
     walletAddress: 'https://wallet.elrond.com',
     apiAddress: 'https://api.elrond.com',
     explorerAddress: 'https://explorer.elrond.com',